Copying bug fix from jana_0.7.4p2.
git-svn-id: https://phys12svn.jlab.org/repos/JANA@2405 c5ed4466-e916-0410-8347-b3263e9c103d

+++ src/plugins/janaview/jv_mainframe.cc
@@ -168,7 +168,7 @@ void jv_mainframe::DoSelectObjectType(Int_t id)
 	// Get factory name and tag
 	string name = nametag;
 	string tag = "";
-	uint32_t pos=name.find(":");
+	size_t pos=name.find(":");
 	if(pos != name.npos){
 		tag = name.substr(pos+1);
 		name = name.substr(0, pos);
